perf: improve FCP/LCP by optimizing CSS and splitting page component
- Replace `@import "tailwindcss"` with `@import "tailwindcss/preflight" layer(base)`
  to eliminate unused Tailwind utility CSS (~128 KiB saved, est. 600ms FCP improvement)
- Convert Tailwind utility classes in FileUpload and about-score to custom CSS
  classes to allow removal of full Tailwind import
- Split `page.tsx` into Server Component with static H1 + `HomeClient` client
  component so the LCP element renders without waiting for JS hydration
- Remove unused font weight 500 from Noto Sans JP to reduce font CSS
  from 375KB to 282KB (~25% reduction, est. 185ms savings)

Closes #304
