+++ webapp/src/app/layout.tsx
@@ -5,7 +5,7 @@ import Sidebar from '@/components/Sidebar'
 import { LanguageProvider } from '@/lib/i18n'
 
 const notoSansJP = Noto_Sans_JP({
-  weight: ['400', '500', '600', '700'],
+  weight: ['400', '600', '700'],
   subsets: ['latin'],
   preload: false,
   variable: '--font-noto-sans-jp',
